requirements.txt
pycryptodome
sympy
requests
argparse